--- packages/babel-generator/test/fixtures/comments/create-parenthesized-expressions/options.json
@@ -1,3 +1,0 @@
-{
-  "parserOpts": {"createParenthesizedExpressions": true}
-}
