--- packages/babel-generator/test/fixtures/comments/create-parenthesized-expressions/output.js
@@ -1,10 +1,0 @@
-// One
-(1);
-/* Two */
-
-(2);
-( // Three
-3);
-(
-/* Four */
-4);
