+++ packages/babel-generator/test/index.js
@@ -492,7 +492,6 @@ suites.forEach(function(testSuite) {
               strictMode: false,
               sourceType: "module",
               sourceMaps: !!task.sourceMap,
-              ...task.options.parserOpts,
             });
             const options = {
               sourceFileName: path.relative(__dirname, actual.loc),
