+++ packages/babel-parser/src/parser/expression.js
@@ -185,28 +185,17 @@ export default class ExpressionParser extends LValParser {
 
       let patternErrorMsg;
       let elementName;
-
-      const unwrap = node => {
-        return node.type === "ParenthesizedExpression"
-          ? unwrap(node.expression)
-          : node;
-      };
-      const maybePattern = unwrap(left);
-      if (maybePattern.type === "ObjectPattern") {
+      if (left.type === "ObjectPattern") {
         patternErrorMsg = "`({a}) = 0` use `({a} = 0)`";
         elementName = "property";
-      } else if (maybePattern.type === "ArrayPattern") {
+      } else if (left.type === "ArrayPattern") {
         patternErrorMsg = "`([a]) = 0` use `([a] = 0)`";
         elementName = "element";
       }
 
-      if (
-        patternErrorMsg &&
-        ((left.extra && left.extra.parenthesized) ||
-          left.type === "ParenthesizedExpression")
-      ) {
+      if (patternErrorMsg && left.extra && left.extra.parenthesized) {
         this.raise(
-          maybePattern.start,
+          left.start,
           `You're trying to assign to a parenthesized expression, eg. instead of ${patternErrorMsg}`,
         );
       }
@@ -320,8 +309,7 @@ export default class ExpressionParser extends LValParser {
         if (
           operator === "**" &&
           left.type === "UnaryExpression" &&
-          (this.options.createParenthesizedExpressions ||
-            !(left.extra && left.extra.parenthesized))
+          !(left.extra && left.extra.parenthesized)
         ) {
           this.raise(
             left.argument.start,
@@ -1167,6 +1155,13 @@ export default class ExpressionParser extends LValParser {
     return this.finishNode(node, type);
   }
 
+  parseParenExpression(): N.Expression {
+    this.expect(tt.parenL);
+    const val = this.parseExpression();
+    this.expect(tt.parenR);
+    return val;
+  }
+
   parseParenAndDistinguishExpression(canBeArrow: boolean): N.Expression {
     const startPos = this.state.start;
     const startLoc = this.state.startLoc;
@@ -1277,16 +1272,10 @@ export default class ExpressionParser extends LValParser {
       val = exprList[0];
     }
 
-    if (!this.options.createParenthesizedExpressions) {
-      this.addExtra(val, "parenthesized", true);
-      this.addExtra(val, "parenStart", startPos);
-      return val;
-    }
+    this.addExtra(val, "parenthesized", true);
+    this.addExtra(val, "parenStart", startPos);
 
-    const parenExpression = this.startNodeAt(startPos, startLoc);
-    parenExpression.expression = val;
-    this.finishNode(parenExpression, "ParenthesizedExpression");
-    return parenExpression;
+    return val;
   }
 
   shouldParseArrow(): boolean {
