+++ packages/babel-parser/src/parser/lval.js
@@ -92,14 +92,6 @@ export default class LValParser extends NodeUtils {
           }
           break;
 
-        case "ParenthesizedExpression":
-          node.expression = this.toAssignable(
-            node.expression,
-            isBinding,
-            contextDescription,
-          );
-          break;
-
         case "MemberExpression":
           if (!isBinding) break;
 
@@ -413,15 +405,6 @@ export default class LValParser extends NodeUtils {
 
       case "RestElement":
         this.checkLVal(expr.argument, isBinding, checkClashes, "rest element");
-        break;
-
-      case "ParenthesizedExpression":
-        this.checkLVal(
-          expr.expression,
-          isBinding,
-          checkClashes,
-          "parenthesized expression",
-        );
         break;
 
       default: {
