--- packages/babel-parser/test/fixtures/comments/basic/create-parenthesized-expressions/input.js
@@ -1,12 +1,0 @@
-// One
-(1);
-
-/* Two */
-(2);
-
-(
-  // Three
-  3
-);
-
-(/* Four */ 4);
