--- packages/babel-parser/test/fixtures/comments/basic/create-parenthesized-expressions/options.json
@@ -1,3 +1,0 @@
-{
-  "createParenthesizedExpressions": true
-}
