--- packages/babel-parser/test/fixtures/core/uncategorised/556/input.js
@@ -1,1 +1,0 @@
-({x}) = {x: 1};
