--- packages/babel-parser/test/fixtures/core/uncategorised/557/input.js
@@ -1,1 +1,0 @@
-("hello");
