--- packages/babel-parser/test/fixtures/core/uncategorised/557/options.json
@@ -1,3 +1,0 @@
-{
-  "createParenthesizedExpressions": true
-}
