--- packages/babel-parser/test/fixtures/core/uncategorised/557/output.json
@@ -1,85 +1,0 @@
-{
-  "type": "File",
-  "start": 0,
-  "end": 10,
-  "loc": {
-    "start": {
-      "line": 1,
-      "column": 0
-    },
-    "end": {
-      "line": 1,
-      "column": 10
-    }
-  },
-  "program": {
-    "type": "Program",
-    "start": 0,
-    "end": 10,
-    "loc": {
-      "start": {
-        "line": 1,
-        "column": 0
-      },
-      "end": {
-        "line": 1,
-        "column": 10
-      }
-    },
-    "sourceType": "script",
-    "interpreter": null,
-    "body": [
-      {
-        "type": "ExpressionStatement",
-        "start": 0,
-        "end": 10,
-        "loc": {
-          "start": {
-            "line": 1,
-            "column": 0
-          },
-          "end": {
-            "line": 1,
-            "column": 10
-          }
-        },
-        "expression": {
-          "type": "ParenthesizedExpression",
-          "start": 0,
-          "end": 9,
-          "loc": {
-            "start": {
-              "line": 1,
-              "column": 0
-            },
-            "end": {
-              "line": 1,
-              "column": 9
-            }
-          },
-          "expression": {
-            "type": "StringLiteral",
-            "start": 1,
-            "end": 8,
-            "loc": {
-              "start": {
-                "line": 1,
-                "column": 1
-              },
-              "end": {
-                "line": 1,
-                "column": 8
-              }
-            },
-            "extra": {
-              "rawValue": "hello",
-              "raw": "\"hello\""
-            },
-            "value": "hello"
-          }
-        }
-      }
-    ],
-    "directives": []
-  }
-}
