--- packages/babel-parser/test/fixtures/core/uncategorised/558/input.js
@@ -1,1 +1,0 @@
-([a]) = []
