--- packages/babel-parser/test/fixtures/core/uncategorised/558/options.json
@@ -1,4 +1,0 @@
-{
-  "createParenthesizedExpressions": true,
-  "throws": "You're trying to assign to a parenthesized expression, eg. instead of `([a]) = 0` use `([a] = 0)` (1:1)"
-}
