--- packages/babel-parser/test/fixtures/es2015/arrow-functions/create-parenthesized-expressions/input.js
@@ -1,1 +1,0 @@
-(foo) => {}
