--- packages/babel-parser/test/fixtures/es2015/arrow-functions/create-parenthesized-expressions/options.json
@@ -1,3 +1,0 @@
-{
-  "createParenthesizedExpressions": true
-}
