--- packages/babel-parser/test/fixtures/es2015/arrow-functions/create-parenthesized-expressions/output.json
@@ -1,104 +1,0 @@
-{
-  "type": "File",
-  "start": 0,
-  "end": 11,
-  "loc": {
-    "start": {
-      "line": 1,
-      "column": 0
-    },
-    "end": {
-      "line": 1,
-      "column": 11
-    }
-  },
-  "program": {
-    "type": "Program",
-    "start": 0,
-    "end": 11,
-    "loc": {
-      "start": {
-        "line": 1,
-        "column": 0
-      },
-      "end": {
-        "line": 1,
-        "column": 11
-      }
-    },
-    "sourceType": "script",
-    "interpreter": null,
-    "body": [
-      {
-        "type": "ExpressionStatement",
-        "start": 0,
-        "end": 11,
-        "loc": {
-          "start": {
-            "line": 1,
-            "column": 0
-          },
-          "end": {
-            "line": 1,
-            "column": 11
-          }
-        },
-        "expression": {
-          "type": "ArrowFunctionExpression",
-          "start": 0,
-          "end": 11,
-          "loc": {
-            "start": {
-              "line": 1,
-              "column": 0
-            },
-            "end": {
-              "line": 1,
-              "column": 11
-            }
-          },
-          "id": null,
-          "generator": false,
-          "async": false,
-          "params": [
-            {
-              "type": "Identifier",
-              "start": 1,
-              "end": 4,
-              "loc": {
-                "start": {
-                  "line": 1,
-                  "column": 1
-                },
-                "end": {
-                  "line": 1,
-                  "column": 4
-                },
-                "identifierName": "foo"
-              },
-              "name": "foo"
-            }
-          ],
-          "body": {
-            "type": "BlockStatement",
-            "start": 9,
-            "end": 11,
-            "loc": {
-              "start": {
-                "line": 1,
-                "column": 9
-              },
-              "end": {
-                "line": 1,
-                "column": 11
-              }
-            },
-            "body": [],
-            "directives": []
-          }
-        }
-      }
-    ],
-    "directives": []
-  }
-}
