--- packages/babel-parser/test/fixtures/es2015/uncategorised/394/input.js
@@ -1,4 +1,0 @@
-var a;
-(a) = {};
-(a.b) = {};
-(a['c']) = {};
