--- packages/babel-parser/test/fixtures/es2015/uncategorised/394/options.json
@@ -1,3 +1,0 @@
-{
-  "createParenthesizedExpressions": true
-}
