--- packages/babel-parser/test/fixtures/es2015/uncategorised/395/input.js
@@ -1,1 +1,0 @@
-([a.a]) => 42
