--- packages/babel-parser/test/fixtures/es2015/uncategorised/395/options.json
@@ -1,4 +1,0 @@
-{
-  "createParenthesizedExpressions": true,
-  "throws": "Invalid left-hand side in arrow function parameters (1:2)"
-}
