--- packages/babel-parser/test/fixtures/es2016/exponentiation-operator/16/input.js
@@ -1,1 +1,0 @@
-(-5 ** 6);
