--- packages/babel-parser/test/fixtures/es2016/exponentiation-operator/16/options.json
@@ -1,4 +1,0 @@
-{
-  "createParenthesizedExpressions": true,
-  "throws": "Illegal expression. Wrap left hand side or entire exponentiation in parentheses. (1:2)"
-}
