+++ packages/babel-parser/test/fixtures/es2016/exponentiation-operator/2/output.json
