+++ packages/babel-parser/test/fixtures/es2016/exponentiation-operator/7/output.json
