+++ packages/babel-parser/test/fixtures/es2016/exponentiation-operator/8/output.json
