+++ packages/babel-parser/test/fixtures/es2016/exponentiation-operator/9/output.json
