+++ packages/babel-parser/typings/babel-parser.d.ts
@@ -78,14 +78,6 @@ export interface ParserOptions {
      * Adds all parsed tokens to a tokens property on the File node.
      */
     tokens?: boolean;
-
-    /**
-     * By default, the parser adds information about parentheses by setting
-     * `extra.parenthesized` to `true` as needed.
-     * When this option is `true` the parser creates `ParenthesizedExpression`
-     * AST nodes instead of using the `extra` property.
-     */
-    createParenthesizedExpressions?: boolean;
 }
 
 export type ParserPlugin =
