+++ packages/babel-types/src/definitions/core.js
@@ -731,16 +731,6 @@ defineType("SequenceExpression", {
   aliases: ["Expression"],
 });
 
-defineType("ParenthesizedExpression", {
-  visitor: ["expression"],
-  aliases: ["Expression", "ExpressionWrapper"],
-  fields: {
-    expression: {
-      validate: assertNodeType("Expression"),
-    },
-  },
-});
-
 defineType("SwitchCase", {
   visitor: ["test", "consequent"],
   fields: {
