+++ packages/babel-types/src/definitions/misc.js
@@ -1,6 +1,16 @@
 // @flow
-import defineType from "./utils";
+import defineType, { assertNodeType } from "./utils";
 
 defineType("Noop", {
   visitor: [],
+});
+
+defineType("ParenthesizedExpression", {
+  visitor: ["expression"],
+  aliases: ["Expression", "ExpressionWrapper"],
+  fields: {
+    expression: {
+      validate: assertNodeType("Expression"),
+    },
+  },
 });
